requirements.txt
fastapi
uvicorn
spotipy
python-dotenv
google-generativeai